requirements.txt
Django==1.11.11
djangorestframework==3.3.3
MySQL-python
